Fix for TPINVTRY-224 - Inconsistent swagger response codes for resource updates.
This fixes https://projects.engineering.redhat.com/browse/TPINVTRY-224

The 400 is actually the proper return code for trying to update read-only attributes,
i.e. a Bad Request. We generate the 400's properly as we catch
the ActionController::UnpermittedParameters error.

Fix was to update the swagger_generate.rake script to include a 400 response
code for collection updates.  This is effectively reflected for
the authentications, endpoints, sources and tasks collections.